refactor: 확장성 용이를 위한 Repository 클래스 추상화 - saveUser 메서드 명세 - findById 메서드 명세 - findByUserId 메서드 명세 - findAll 메서드 명세
